📦 NEW: Move route pattern funcs to util
